requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
pyarrow>=15.0
PyYAML>=6.0
google-cloud-bigquery>=3.25
google-cloud-storage>=2.18
